# specs/genesys-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "With the PureCloud Platform API, you can control all aspects of your PureCloud environment. With the APIs you can access the system configuration, manage conversations and more.",
    "version": "v2",
    "title": "PureCloud Platform API",
    "termsOfService": "https://help.genesys.cloud/articles/terms-and-conditions/",
    "contact": {
      "name": "PureCloud Developer Evangelists",
      "url": "https://developer.genesys.cloud/",
      "email": "DeveloperEvangelists@genesys.com"
    },
    "license": {
      "name": "UNLICENSED",
      "url": "https://help.genesys.cloud/articles/terms-and-conditions/"
    }
  },
  "host": "api.mypurecloud.com",
  "basePath": null,
  "paths": {
...
    "/api/v2/authorization/divisions": {
      "get": {
        "operationId": "getAuthorizationDivisions",
        "summary": "Retrieve a list of all divisions defined for the organization",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "schema": null
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "type": "integer",
            "schema": null
          }
        ]
      }
    },
...
    "/api/v2/users": {
      "get": {
        "operationId": "getUsers",
        "summary": "Get the list of available users.",
        "parameters": [
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "type": "integer",
            "schema": null
          },
          {
            "name": "pageNumber",
            "in": "query",
            "required": false,
            "type": "integer",
            "schema": null
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "array",
            "schema": null
          },
          {
            "name": "jabberId",
            "in": "query",
            "required": false,
            "type": "array",
            "schema": null
          },
          {
            "name": "sortOrder",
            "in": "query",
            "required": false,
            "type": "string",
            "schema": null
          },
          {
            "name": "expand",
            "in": "query",
            "required": false,
            "type": "array",
            "schema": null
          },
          {
            "name": "integrationPresenceSource",
            "in": "query",
            "required": false,
            "type": "string",
            "schema": null
          },
          {
            "name": "userCustomAttributeSchemaIds",
            "in": "query",
            "required": false,
            "type": "array",
            "schema": null
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "type": "string",
            "schema": null
          }
        ]
      },
      "post": {
        "operationId": "postUsers",
        "summary": "Create user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": null,
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ]
      }
    },
...
    "/api/v2/users/{userId}": {
      "get": {
        "operationId": "getUser",
        "summary": "Get user.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          },
          {
            "name": "expand",
            "in": "query",
            "required": false,
            "type": "array",
            "schema": null
          },
          {
            "name": "integrationPresenceSource",
            "in": "query",
            "required": false,
            "type": "string",
            "schema": null
          },
          {
            "name": "userCustomAttributeSchemaIds",
            "in": "query",
            "required": false,
            "type": "array",
            "schema": null
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "type": "string",
            "schema": null
          }
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "Delete user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      },
      "patch": {
        "operationId": "patchUser",
        "summary": "Update user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": null,
            "schema": {
              "$ref": "#/definitions/UpdateUser"
            }
          }
        ]
      }
    },
...
    "/api/v2/users/{userId}/state": {
      "get": {
        "operationId": "getUserState",
        "summary": "Get user state information.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      },
      "put": {
        "operationId": "putUserState",
        "summary": "Update user state information.",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "type": null,
            "schema": {
              "$ref": "#/definitions/UserState"
            }
          }
        ]
      }
    },
...
    "/api/v2/voicemail/userpolicies/{userId}": {
      "get": {
        "operationId": "getVoicemailUserpolicy",
        "summary": "Get the voicemail policy for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      },
      "patch": {
        "operationId": "patchVoicemailUserpolicy",
        "summary": "Update the voicemail policy for a user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "schema": null
          }
        ]
      }
    }
  }
}